fix: installing latest docker-compose (#15)
* fix: installing latest docker-compose

* Update README.md

* feat(bootstrap): add docker certs

* feat(bootstrap): add vm.startup script

* feat(bootstrap): add docker certs

* fix(vm.startup): break on sucess

* Update README.md

* fix(bootstrap): filemode for Mac

* fix(vm.startup): add PATH and uname check

* Update README.md

* fix(startup): remove log

+++ bootstrap.sh
@@ -3,8 +3,11 @@
 set +e
 
 sed="sed"
+windows=
 if [ $(uname -s) == "Darwin" ]; then
   sed="gsed"
+else
+  windows=1
 fi
 
 echo =================================
@@ -19,10 +22,10 @@ swapfile=${SWAPFILE:-}
 # get username from env or prompt
 username=$VAGRANT_USERNAME
 if [ -z "$VAGRANT_USERNAME" ]; then
-echo -n "> Please enter default vagrant user name [vagrant]:"
-read input
-username=${input:-vagrant}
-echo "VAGRANT_USERNAME=$username">> .env
+  echo -n "> Please enter default vagrant user name [vagrant]:"
+  read input
+  username=${input:-vagrant}
+  echo "VAGRANT_USERNAME=$username">> .env
 fi
 
 machine_name=${NAME:-linuxdev}
@@ -34,7 +37,7 @@ vagrant up
 # create ssh config file
 SSH_CONFIG="./ssh.config"
 if [ -z "$(grep vagrant $SSH_CONFIG)" ]; then
-vagrant ssh-config >> $SSH_CONFIG
+  vagrant ssh-config >> $SSH_CONFIG
 fi
 
 # create user with UID 1000
@@ -46,43 +49,43 @@ vagrant_uid=$($ssh id -u vagrant 2>/dev/null)
 
 set -e
 
-if [[ $vagrant_uid == 1000 || ($exists != "" && $exists != 1000) ]]; then
+if [ $vagrant_uid == 1000 ] || [ "$exists" != "" && "$exists" != 1000 ]; then
   echo switching is required, remove $username and try again
 fi
-if [[ -z "$vagrant_uid" ]]; then
+if [ -z "$vagrant_uid" ]; then
   echo ssh connection looks like failed
   exit -1;
 fi
 $ssh sudo cp -a /home/vagrant/.ssh /root/
 $ssh sudo chown -R root:root /root/.ssh
 
-if [[ -z "$(grep root $SSH_CONFIG.root)" ]]; then
+if [ -z "$(grep root $SSH_CONFIG.root)" ]; then
   $sed -e '0,/vagrant/{s/vagrant/root/}' -e '0,/default/{s/default/root/}' $SSH_CONFIG >> $SSH_CONFIG.root
 fi
 #### user root
 ssh="ssh -F $SSH_CONFIG.root root"
 
 
-if [[ -z "$exists" ]]; then
-echo "user $username not found"
+if [ -z "$exists" ]; then
+  echo "user $username not found"
   $ssh << EOSSH
 echo ---------------------
 echo "creating $username"
 vagrant_uid=\$(id -u vagrant)
-if [[ \$vagrant_uid == 1000 ]]; then
+if [ \$vagrant_uid == 1000 ]; then
   pkill -U 1000
   usermod -u 1002 vagrant
   groupmod -g 1002 vagrant
 fi
 chown -R vagrant:vagrant /home/vagrant
 useradd $username -u 1000 --create-home
-if [ ! -d "/home/$username/.ssh" ]; then
+if ! [ -d "/home/$username/.ssh" ]; then
   cp -a /home/vagrant/.ssh /home/$username/
   chown -R $username:$username /home/$username/.ssh
 fi
 grep $username /etc/passwd
 EOSSH
-echo ---------------------
+  echo ---------------------
 fi
 
 # Adding $username to Sudoer 
@@ -96,7 +99,7 @@ usermod -aG docker $username
 
 if [ $swapfile ]; then
   echo Found SWAPFILE config
-  if [[ ! -f /swapfile ]]; then
+  if ! [ -f /swapfile ]; then
     echo "-----
 Creating swapfile"
     dd if=/dev/zero of=/swapfile bs=1M count=1024 oflag=append conv=notrunc
@@ -106,7 +109,7 @@ Creating swapfile"
   echo "-----
 Adding swapfile"
   sudo swapon /swapfile
-  if [[ -z "\$(grep swapfile -w /etc/fstab)" ]]; then
+  if [ -z "\$(grep swapfile -w /etc/fstab)" ]; then
     echo "/swapfile swap swap defaults 0 0" >> /etc/fstab
   fi
   mount -a
@@ -114,7 +117,7 @@ fi
 swapon --show
 free -h
 
-if [[ ! -f /dummy ]]; then
+if ! [ -f /dummy ]; then
   echo "-----
 Expanding actual size for ${expand_disk_size}GB"
   let "blockSize = $expand_disk_size * 1024"
@@ -123,11 +126,23 @@ Expanding actual size for ${expand_disk_size}GB"
   dd if=/dev/zero of=/dummy bs=1M count=\$blockSize oflag=append conv=notrunc
 fi
 
+if [ -z "\$(crontab -l|grep "${machine_name}.startup.sh")" ]; then
+  echo "-----
+Adding startup script to crontab"
+  cp /vagrant/vm.startup.sh /root/${machine_name}.startup.sh && \
+  chmod +x /root/${machine_name}.startup.sh && \
+  crontab -l | { cat; echo "@reboot /root/${machine_name}.startup.sh"; } | crontab -
+else
+  echo "-----
+crontab scripts:"
+fi
+  crontab -l
+
 EOSSH
 
 echo ---------------------
-if [[ -z $(grep $machine_name ~/.ssh/config) ]]; then
-  echo adding ssh config for $machine_name
+if [ -z "$(grep $machine_name ~/.ssh/config)" ]; then
+  echo Adding ssh config for $machine_name
   $sed -e "0,/vagrant/{s/vagrant/$username/}" -e "0,/default/{s/default/$machine_name/}" $SSH_CONFIG >> ~/.ssh/config
 else
   echo $machine_name entry found in ~/.ssh/config. Please double check if Port is correct:
@@ -145,14 +160,31 @@ zsh \
 vim-gtk \
 python3-pip \
 dnsutils \
+pass gnupg2 \
 -y
-wget https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
-sh install.sh --unattended
-rm -f install.sh*
-sudo chsh -s /bin/zsh $username
-echo "-----\nInstalling docker-compose...."
-sudo apt install docker-compose -y
+
+if [ -f ~/.oh-my-zsh/oh-my-zsh.sh ]; then
+  echo "-----\noh my zsh is aleady installed"
+else
+  echo "-----\nInstalling oh my zsh...."
+  wget https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
+  sh install.sh --unattended
+  rm -f install.sh*
+  sudo chsh -s /bin/zsh $username
+fi
+
+if [ -f /usr/local/bin/docker-compose ]; then
+  echo "-----\ndocker-compose aleady exists"
+else
+  echo "-----\nInstalling docker-compose...."
+  docker_compose_url=https://github.com\$(wget -q -O - https://github.com/docker/compose/releases/latest | sed -n 's/.*href="\([^"]*\).*/\1/p'|grep "\$(uname -s)-\$(uname -m)$")
+  echo docker_compose_url: \$docker_compose_url
+  sudo wget \$docker_compose_url -O /usr/local/bin/docker-compose
+  sudo chmod +x /usr/local/bin/docker-compose
+  sudo pip3 install requests --upgrade
+fi
 docker-compose --version
+
 echo "-----\nConfiguring samba"
 mkdir -p ~/Projects
 mkdir -p samba
@@ -165,10 +197,10 @@ chmod +x adduser
 ./adduser \$USER
 
 if [ -f ~/.ssh/id_rsa ]; then
-echo "-----\nssh key aleady exists"
+  echo "-----\nssh key aleady exists"
 else
-echo "-----\nGenerating ssh key"
-ssh-keygen -b 2048 -t rsa -f ~/.ssh/id_rsa -q -N ""
+  echo "-----\nGenerating ssh key"
+  ssh-keygen -b 2048 -t rsa -f ~/.ssh/id_rsa -q -N ""
 fi
 echo "Paste the public key below into Github or else"
 echo ---------------------
@@ -177,7 +209,7 @@ echo ---------------------
 
 if [ -f /dummy ]; then
   filesize=\$(stat -c%s "/dummy")
-  if (( filesize > 1 )); then
+  if [ "\$filesize" ] && [ "\$filesize" != "0" ]; then
     echo \$filesize was larger than 1, removing /dummy
     sudo rm /dummy
     sudo touch /dummy
@@ -185,7 +217,29 @@ if [ -f /dummy ]; then
 fi
 EOSSH
 
-echo "
+if ! [ -d ~/.docker/certs.$machine_name ]; then
+  echo "--------
+Creating Docker certs"
+  ssh $machine_name /vagrant/create_docker_certs.sh
+  mkdir -p ~/.docker/certs.$machine_name
+  cp ./certs/*.pem ~/.docker/certs.$machine_name/
+  ssh $machine_name /vagrant/config_docker_certs.sh
+  echo "export DOCKER_CERT_PATH=~/.docker/certs.$machine_name
+export DOCKER_HOST=192.168.99.123
+export DOCKER_TLS_VERIFY=1
+" >> ~/.bashrc
+touch ~/.bash_profile
+fi
+if [ "$windows" ] && ! [ -f ~/docker_env.bat ]; then
+  echo "@echo off
+set DOCKER_CERT_PATH=%userprofile%\.docker\certs.$machine_name
+set DOCKER_HOST=192.168.99.123
+set DOCKER_TLS_VERIFY=1
+" > ~/docker_env.bat
+fi
+
+echo "----------------------
+
 Congrats!!!
 
 You can now ssh into the machine by
